Original change: https://android-review.googlesource.com/c/platform/packages/modules/NetworkStack/+/2606684

+++ src/android/net/dhcp6/Dhcp6AdvertisePacket.java
@@ -34,7 +34,7 @@ public class Dhcp6AdvertisePacket extends Dhcp6Packet {
      */
     Dhcp6AdvertisePacket(int transId, @NonNull final byte[] clientDuid,
             @NonNull final byte[] serverDuid, final byte[] iapd) {
-        super(transId, (short) 0 /* secs */, clientDuid, serverDuid, iapd);
+        super(transId, 0 /* elapsedTime */, clientDuid, serverDuid, iapd);
     }
 
     /**
